Rename `janet.py` to `app.py`
